+++ CHANGELOG.md
@@ -14,6 +14,7 @@ _**Note:** Yet to be released breaking changes appear here._
   The method now accepts a single parameter, mainly to minimize the need to pass many default values.
 - The `AbstractGraph.getPlugin` method now explicitly returns `undefined` when a plugin is not found.  
   TypeScript users must update their code to handle the `undefined` case when calling this method.
+- The `Dictionary` class has been removed. Use the `Map` class instead, which is a standard JavaScript feature.
 
 ## 0.20.0